+++ src/ui/tab_strip.rs
@@ -1331,7 +1331,7 @@ impl Tty7App {
             let rename_input = self
                 .renaming
                 .as_ref()
-                .filter(|r| r.index == i)
+                .filter(|r| r.tab == tab.tree_id.get())
                 .map(|r| r.input.clone());
             let label_region = match rename_input {
                 Some(input) => div()